postureChecker/scripts/requirements.txt
bleak
memfault-cli
pyserial-asyncio
strictdoc
